refactor(app): split App.New into initialization phases
